elasticsearch 설정 추가

+++ src/main/java/oridungjeol/duckhang/payment/domain/Payment.java
@@ -10,6 +10,6 @@ public class Payment {
     private final String orderId;     // 주문 고유 ID
     private final String paymentKey;  // 결제 키
     private final int amount;         // 결제 금액
-    private final int boardId;        // 게시글의 board_id (외래키)
+    private final Long boardId;        // 게시글의 board_id (외래키)
     private final String type;        // 게시글 타입 (SELL, RENTAL 등)
 }
